Move aws helper functions to seperate pkg/aws
